job closing and more documentation
I wrote a bunch of documentation for how to implement job closing on the frontend. Shouldn't be super complex, but essentially a company can close or decide their job. If they close it they just aren't accepting any more applicants but still can make a decision on an applicant, but if they decide it, all non-accepted applicants will be auto-rejected and the company will not be able to update the job or it's applicants anymore.