Added allApps
this is for the page that shows a user all of their applications. I left some basic documentation about my backend code, but it's really simple and basically just returns a list of applications a user has made. Whoever is making the page, please look in userAllApps/app.js for documentation and put html code in userAllApps/userAllApps.html